Add explicit cast in vector2d::getAngleWith to avoid warnings when using type irr::s64.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@4122 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ include/vector2d.h
@@ -237,7 +237,7 @@ public:
 	\return Returns a value between 0 and 90. */
 	inline f64 getAngleWith(const vector2d<T>& b) const
 	{
-		f64 tmp = X*b.X + Y*b.Y;
+		f64 tmp = (f64)(X*b.X + Y*b.Y);
 
 		if (tmp == 0.0)
 			return 90.0;
